v0.7.2
[Enhancement] Fixed issues in the code which caused warnings in MSVC and clang compilers.

[Enhancement] Fixed errors in `get_heuristics_list` where for certain heuristics mode in older cuDNN versions, the heuristics list might be incorrect.

[Bug fixes] Fixed several test cases failing on unsupported GPUs to exit gracefully.

[Samples] Added a sample to showcase fp8 convolution forward in Nvidia Hopper GPUs. The sample also showcases post convolution book-keeping operations such as scaling and absolute maximum reduction.

[Samples] Added a sample which converts fp16 tensor to fp8 and performs transpose and absolute maximum reduction.

[Samples] Added a sample to demonstrate Max pooling operation including tensor index dump, necessary to speed up the backward pass.

[Samples] Added a sample to showcase the backward pooling operation.

+++ include/cudnn_frontend_VariantPack.h
@@ -90,7 +90,7 @@ class VariantPackBuilder_v8 {
     //! Set dataPointers for the VariantPack_v8
     auto
     setDataPointers(int64_t num_ptr, void **ptrs) -> VariantPackBuilder_v8 & {
-        m_variant_pack.data_pointers.reserve(num_ptr);
+        m_variant_pack.data_pointers.reserve(static_cast<size_t>(num_ptr));
         std::copy(ptrs, ptrs + num_ptr, std::back_inserter(m_variant_pack.data_pointers));
         m_variant_pack.num_ptrs = num_ptr;
         return *this;
@@ -103,7 +103,7 @@ class VariantPackBuilder_v8 {
 
     auto
     setUids(int64_t num_uids, int64_t *uid) -> VariantPackBuilder_v8 & {
-        m_variant_pack.uid.reserve(num_uids);
+        m_variant_pack.uid.reserve(static_cast<size_t>(num_uids));
         std::copy(uid, uid + num_uids, std::back_inserter(m_variant_pack.uid));
         return *this;
     }
@@ -111,8 +111,8 @@ class VariantPackBuilder_v8 {
     auto
     setDataPointers(std::set<std::pair<uint64_t, void *>> const &data_pointers) -> VariantPackBuilder_v8 & {
         m_variant_pack.num_ptrs = data_pointers.size();
-        m_variant_pack.uid.reserve(m_variant_pack.num_ptrs);
-        m_variant_pack.data_pointers.reserve(m_variant_pack.num_ptrs);
+        m_variant_pack.uid.reserve(static_cast<size_t>(m_variant_pack.num_ptrs));
+        m_variant_pack.data_pointers.reserve(static_cast<size_t>(m_variant_pack.num_ptrs));
         for (auto &data_pointer : data_pointers) {
             m_variant_pack.uid.push_back(data_pointer.first);
             m_variant_pack.data_pointers.push_back(data_pointer.second);
